+++ src/ui/AnalyzerPopup.hpp
@@ -3,6 +3,8 @@
 #include <Geode/utils/file.hpp>
 #include <Geode/utils/async.hpp>
 #include "../Replay.hpp"
+#include "tabs/Stats/StatsTab.hpp"
+#include "tabs/Analysis/AnalysisTab.hpp"
 
 using namespace geode::prelude;
 
@@ -10,14 +12,17 @@ class AnalyzerPopup : public Popup {
 protected:
     bool init(float width, float height);
     arc::TaskHandle<void> m_pickHandle;
-    CCNode* m_statsContainer = nullptr;
+    StatsTab* m_statsTab = nullptr;
+    AnalysisTab* m_analysisTab = nullptr;
+    CCMenuItemSpriteExtra* m_statsTabBtn = nullptr;
+    CCMenuItemSpriteExtra* m_analysisTabBtn = nullptr;
     CCNode* m_loadingContainer = nullptr;
     CCLabelBMFont* m_errorLabel = nullptr;
-    CCNode* createMetricCard(std::string const& label, std::string const& value, ccColor3B color, float delay);
-    void populateStats(Replay const& replay);
     void showLoading();
     void hideLoading();
     void showError(std::string const& message);
+    void switchTab(bool showAnalysis);
+    void onSwitchTab(CCObject* sender);
 
 public:
     static AnalyzerPopup* create();